dhcp 12.7-3

+++ services/dhcp_server/src/dhcp_s_server.cpp
@@ -1189,7 +1189,8 @@ static int OnReceivedRequest(PDhcpServerContext ctx, PDhcpMsgInfo received, PDhc
         AddAddressOption(reply, ROUTER_OPTION, srvIns->addressPool.gateway);
     }
     /* Add option43 as "OPEN_HARMONY". */
-    DhcpOption optVendorInfo = {VENDOR_SPECIFIC_INFO_OPTION, strlen("OPEN_HARMONY"), "OPEN_HARMONY"};
+    DhcpOption optVendorInfo = {
+        VENDOR_SPECIFIC_INFO_OPTION, static_cast<uint8_t>(strlen("OPEN_HARMONY")), "OPEN_HARMONY"};
     PushBackOption(&reply->options, &optVendorInfo);
     reply->packet.yiaddr = bindingIp;
     return REPLY_ACK;
